Stop storing blank values in applicant data.
If updates are empty, they
are not put into the applicant data json, but metadata is still
updated. Arrays (enumerator and multi select questions) with empty
updates get deleted, but the metadata is still updated.

Questions are "answered" if any of their scalars have a value in
applicant data. There might be weird behavior with old applicant data
that is storing blank values, but blank values shouldn't exist in the
future.

Errors for multi-field questions (address and name) for each of their
fields are present if the question is answered, not just if the field is
answered.